# docs/openapi.yaml
openapi: 3.1.0
info:
  title: DVinyl API
  version: 1.0.0
  license:
    name: MIT
    identifier: MIT
  description: |
    Mobile/native client API for a DVinyl instance, covering authentication,
    collection, wishlist and sharing management, the full item lifecycle, plugin
    metadata, instance administration, and account settings.

    All routes are mounted under `/api/v1`. Requests and responses are JSON
    (`application/json`), except the file-upload endpoints (`multipart/form-data`),
    the share-link QR endpoint (`image/png`) and the import/export endpoints, whose
    exports answer CSV, ZIP or JSON file downloads.

    Authentication is a JWT Bearer access token obtained from `POST /auth/login`.
    Access tokens live 15 minutes (`expiresIn: 900`); use the rotating refresh
    token from `POST /auth/refresh`. Collection-scoped routes are additionally
    gated on the caller's role in the collection named in the path
    (`viewer` < `editor` < `admin`); instance admins pass every collection role.
    Admin routes require `isAdmin` on the account.

    Error envelope: every failed response is `{ "success": false, "error": "..." }`
    with the HTTP status appropriate to the cause. Success responses are
    resource-shaped (e.g. `{ "item": ... }`, `{ "items": [...], "page": ... }`).

    On an instance that has not completed first-run setup, every `/api/v1`
    request returns `503` with `{ "success": false, "error": "Instance not set up yet" }`.
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /admin/instance-settings:
    get:
      tags: [Admin]
      operationId: getInstanceSettings
      summary: Get instance-wide settings
      description: Requires an instance-admin account.
      responses:
        '200':
          description: Current instance settings.
          content:
            application/json:
              schema:
                type: object
                required: [settings]
                properties:
                  settings:
                    $ref: '#/components/schemas/InstanceSettings'
              example:
                settings:
                  allowMemberCollectionCreation: false
                  maxCollectionsPerUser: 1
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    patch:
      tags: [Admin]
      operationId: updateInstanceSettings
      summary: Update instance-wide settings
      description: |
        Requires an instance-admin account. Body must be JSON. Only recognized
        keys are applied; `maxCollectionsPerUser` is clamped to 1-100.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstanceSettingsUpdateRequest'
            example:
              allowMemberCollectionCreation: true
              maxCollectionsPerUser: 3
      responses:
        '200':
          description: Updated settings.
          content:
            application/json:
              schema:
                type: object
                required: [settings]
                properties:
                  settings:
                    $ref: '#/components/schemas/InstanceSettings'
              example:
                settings:
                  allowMemberCollectionCreation: true
                  maxCollectionsPerUser: 3
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /admin/users:
    get:
      tags: [Admin]
      operationId: listUsers
      summary: List all users
      description: Requires an instance-admin account. Sorted by last change, newest first.
      responses:
        '200':
          description: All users.
          content:
            application/json:
              schema:
                type: object
                required: [users]
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminUser'
              example:
                users:
                  - id: 64f1c2a9e13b2a0012ab34cd
                    username: siebe
                    email: siebe@example.com
                    isAdmin: true
                    lastChange: '2026-09-01T12:00:00.000Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    post:
      tags: [Admin]
      operationId: createUser
      summary: Create a user
      description: |
        Requires an instance-admin account. Body must be JSON. Creates a
        non-admin user with a server-generated random password returned as
        `generatedPassword`. 409 on a username/email collision.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
            example:
              username: guest
              email: guest@example.com
      responses:
        '201':
          description: User created.
          content:
            application/json:
              schema:
                type: object
                required: [user, generatedPassword]
                properties:
                  user:
                    $ref: '#/components/schemas/AdminUser'
                  generatedPassword: { type: string }
              example:
                user:
                  id: 64f1c2a9e13b2a0012ab34c1
                  username: guest
                  email: guest@example.com
                  isAdmin: false
                generatedPassword: 'k7@Rn2pQ!aZ4'
        '400':
          description: Missing fields or invalid username/email.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                missing:
                  value: { success: false, error: 'username and email are required' }
                invalid:
                  value: { success: false, error: 'Invalid username or email' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Username or email already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example: { success: false, error: 'A user with that username or email already exists' }
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /admin/users/{userId}/reset-password:
    post:
      tags: [Admin]
      operationId: resetUserPassword
      summary: Reset a user's password
      description: |
        Requires an instance-admin account. Generates and returns a new random
        password. Instance admins may reset their own password but not another
        admin's (403).
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: New password generated and returned.
          content:
            application/json:
              schema:
                type: object
                required: [generatedPassword]
                properties:
                  generatedPassword: { type: string }
              example: { generatedPassword: 'g8#Lm2vQ!xW1' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Target is another instance admin.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example: { success: false, error: "Cannot reset another admin's password" }
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example: { success: false, error: 'User not found' }
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /admin/users/{userId}:
    delete:
      tags: [Admin]
      operationId: deleteUser
      summary: Delete a user
      description: |
        Requires an instance-admin account. Refuses to delete your own account
        (400) or another instance admin (403). Also removes the user from every
        collection's membership list.
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User deleted.
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success: { type: boolean, enum: [true] }
              example: { success: true }
        '400':
          description: Attempt to delete your own account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example: { success: false, error: 'Cannot delete your own account' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Target is another instance admin.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example: { success: false, error: 'Cannot delete another admin' }
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example: { success: false, error: 'User not found' }
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /admin/blocked-ips:
    get:
      tags: [Admin]
      operationId: listBlockedIps
      summary: List blocked IP addresses
      description: Requires an instance-admin account. Newest first.
      responses:
        '200':
          description: Blocked IPs.
          content:
            application/json:
              schema:
                type: object
                required: [blockedIps]
                properties:
                  blockedIps:
                    type: array
                    items:
                      $ref: '#/components/schemas/BlockedIp'
              example:
                blockedIps:
                  - id: 64f1c2a9e13b2a0012ab3600
                    ip: 203.0.113.7
                    createdAt: '2026-09-10T08:30:00.000Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    post:
      tags: [Admin]
      operationId: blockIp
      summary: Block an IP address
      description: |
        Requires an instance-admin account. Body must be JSON. Blocking an
        already-blocked IP returns 200 with the existing entry; otherwise 201.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlockedIpRequest'
            example:
              ip: 203.0.113.7
      responses:
        '200':
          description: IP was already blocked.
          content:
            application/json:
              schema:
                type: object
                required: [blockedIp]
                properties:
                  blockedIp:
                    $ref: '#/components/schemas/BlockedIp'
              example:
                blockedIp:
                  id: 64f1c2a9e13b2a0012ab3600
                  ip: 203.0.113.7
                  createdAt: '2026-09-10T08:30:00.000Z'
        '201':
          description: IP blocked.
          content:
            application/json:
              schema:
                type: object
                required: [blockedIp]
                properties:
                  blockedIp:
                    $ref: '#/components/schemas/BlockedIp'
              example:
                blockedIp:
                  id: 64f1c2a9e13b2a0012ab3601
                  ip: 198.51.100.4
                  createdAt: '2026-09-11T09:15:00.000Z'
        '400':
          description: ip missing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example: { success: false, error: 'ip is required' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /admin/blocked-ips/{id}:
    delete:
      tags: [Admin]
      operationId: unblockIp
      summary: Unblock an IP address
      description: Requires an instance-admin account.
      parameters:
        - name: id
          in: path
          required: true
          description: Blocked-IP document id.
          schema: { type: string }
      responses:
        '200':
          description: IP unblocked.
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success: { type: boolean, enum: [true] }
              example: { success: true }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example: { success: false, error: 'Not found' }
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /admin/login-logs:
    get:
      tags: [Admin]
      operationId: listLoginLogs
      summary: List recent login attempts
      description: Requires an instance-admin account. Newest first, capped at 200.
      parameters:
        - name: limit
          in: query
          required: false
          description: Maximum entries to return (1-200).
          schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
      responses:
        '200':
          description: Login log entries.
          content:
            application/json:
              schema:
                type: object
                required: [logs]
                properties:
                  logs:
                    type: array
                    items:
                      $ref: '#/components/schemas/LoginLog'
              example:
                logs:
                  - id: 64f1c2a9e13b2a0012ab3700
                    username: siebe
                    email: siebe@example.com
                    ip: 198.51.100.4
                    country: Belgium
                    city: Ghent
                    userAgent: Mozilla/5.0 ...
                    status: success
                    timestamp: '2026-09-11T09:15:00.000Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    delete:
      tags: [Admin]
      operationId: deleteLoginLogs
      summary: Delete the most recent login log entries
      description: |
        Requires an instance-admin account. Deletes the `count` newest entries.
        `count` is required, must be at least 1, and cannot exceed 10000.
      parameters:
        - name: count
          in: query
          required: true
          description: Number of newest entries to delete (1-10000).
          schema: { type: integer, minimum: 1, maximum: 10000 }
      responses:
        '200':
          description: Entries deleted.
          content:
            application/json:
              schema:
                type: object
                required: [deleted]
                properties:
                  deleted: { type: integer }
              example: { deleted: 20 }
        '400':
          description: Missing, non-positive, or over-cap count.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                required:
                  value: { success: false, error: 'count is required and must be >= 1' }
                tooHigh:
                  value: { success: false, error: 'count must be <= 10000' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from `POST /auth/login` (15-minute lifetime).
  parameters:
...
    UserId:
      name: userId
      in: path
      required: true
      description: User id (Mongo ObjectId).
      schema: { type: string }
...
  responses:
...
    Unauthorized:
      description: Missing, invalid or expired Bearer token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missing:
              value: { success: false, error: 'Missing bearer token' }
            invalid:
              value: { success: false, error: 'Invalid or expired token' }
    Forbidden:
      description: Authenticated, but lacking the required role or instance-admin flag.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example: { success: false, error: 'Forbidden' }
...
    InternalServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example: { success: false, error: 'Internal server error' }
    ServiceUnavailable:
      description: The instance has not completed first-run setup; all `/api/v1` routes return this.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example: { success: false, error: 'Instance not set up yet' }
  schemas:
    Error:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, enum: [false] }
        error: { type: string }
        code:
          type: string
          description: Optional machine-readable code; not used by every handler.
      example: { success: false, error: 'Invalid email or password' }
...
    AdminUser:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        email: { type: string, format: email }
        isAdmin: { type: boolean }
        lastChange: { type: string, format: date-time }
...
    BlockedIp:
      type: object
      required: [id, ip]
      properties:
        id: { type: string }
        ip: { type: string }
        createdAt: { type: string, format: date-time }
    LoginLog:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        email: { type: string }
        ip: { type: string }
        country: { type: string }
        city: { type: string }
        userAgent: { type: string }
        status: { type: string }
        timestamp: { type: string, format: date-time }
    InstanceSettings:
      type: object
      required: [allowMemberCollectionCreation, maxCollectionsPerUser]
      properties:
        allowMemberCollectionCreation: { type: boolean }
        maxCollectionsPerUser: { type: integer, minimum: 1, maximum: 100 }
...
    UserCreateRequest:
      type: object
      required: [username, email]
      properties:
        username: { type: string }
        email: { type: string, format: email }
    BlockedIpRequest:
      type: object
      required: [ip]
      properties:
        ip: { type: string }
    InstanceSettingsUpdateRequest:
      type: object
      properties:
        allowMemberCollectionCreation: { type: boolean }
        maxCollectionsPerUser:
          type: integer
          minimum: 1
          maximum: 100
          description: Clamped to 1-100 server-side.
...